fix: remove helper and create internal sign into signer
