--- packages/network/src/signer/signers/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './signature-handler';
