--- packages/network/src/signer/signers/helpers/signature-handler/index.ts
@@ -1,1 +1,0 @@
-export * from './signature-handler';
